[third] Bootstrap: Slightly lower threshold for widest responsive layout
...from 1200 to 1180 pixels.

The monitor I use for Humbug is exactly 1200px wide.  With the scrollbar I come
in just under the original Bootstrap threshold, so I get a scrunched-up nav
sidebar next to a bunch of empty space.

It's annoying to do this in our own CSS because we basically have to duplicate
the whole @media block to make everything fit together.

I don't love editing third-party files like this, but if it gets reverted by a
later update, the consequences are minimal.  If we have important hacks like
this (or just a lot of them), we should decide on a better way to manage them.
For now I have just tagged the commit subject with "[third]".

(imported from commit ef3022b7eb0fdfc8862083bdbb1fb805fbeba2c7)