Disable autocomplete on the search box.
Documentation variously says to disable on the form or on the input box, e.g.:

https://developer.mozilla.org/en-US/docs/How_to_Turn_Off_Form_Autocompletion
http://msdn.microsoft.com/en-us/library/ms533486%28VS.85%29.aspx

However, at least in Chrome, disabling on the form produces an ugly
"This webpage has disabled autocomplete" message if you try to arrow
down on an input box as if there were suggestions, so I've just put it
on the input element. This still works on recent versions of Safari
and Firefox.

(imported from commit b47f438d6e1d930d7d6871070858c97c6479a0ae)