Provide a better link for changing Gravatar
For users logged in to Gravatar, this goes directly to the page where you
change your avatar.  For those not logged in, it's a more directed path to
logging in or signing up, compared to the gravatar.com front page.  However
it's less clear why you are signing up for a new thing, so maybe that's bad.

Also this link uses HTTPS; it seems we can't get that without the 'en.' prefix.

(imported from commit e03f97ef5ed2ff6593a9b349176da5fd056c4514)