Remove link to change email address
We never implemented this; the link goes to a 404.  And it's unlikely that we
will get to it soon, because we currently use email address as the unique
identifier for a user.

Since it can't be changed, we remove it from the 'Settings' tab entirely.
It's still visible in the sidebar, anyway.

(imported from commit a18c0726fb67baf04429a90e36a383cacd03ad8a)