Enforce 10000 char limit on message box
Add maxlength attribute to the textarea which displays a tool tip (in Chrome)
for the user, telling them they are over the limit, so that the error is caught
before the send_message step (which is too late for Create and Send).

Fixes trac ticket #166.

(imported from commit 36b211a17f42dc8345e2235b4b4ca74c8f4e86c4)