Move all of the "alerts" into a row in our navbar.
This is literally a pure "move HTML" operation; no actual code
changes.

This will make them always visible, even when we're in skinny mode
(and will also make more room for our list-of-streams-and-people
sidebar on the left.)

This does have the side-effect that the errors cover up the rest of
the application (in particular, the floating narrowbar). I don't view
this as the end of the world, though the more robust solution would be
for navbar-spacer to grow when the navbar does, and for everything
else to be correspondingly pushed down.

You can visualize an extreme version of what this looks like with
   $(".alert").show()
in the Javascript console.

This fixes Trac #720.

(imported from commit 7a3b12aebcca389465bd1e5852cef5d08fafe3e7)